requirements.txt
torch
torchvision
scikit-learn
nltk
joblib
pillow
numpy
nbformat
